requirements.txt
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.1
idna==2.10
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
scipy==1.6.3
six==1.15.0
Steam-Market-Py==0.0.2
urllib3==1.26.4
